refactor(sections-editor): dedupe basename/extension helpers between image and file fields
