+++ apps/mesh/src/web/components/sections-editor/fields/file-field.tsx
@@ -15,6 +15,7 @@ import { useFilePickerUpload } from "@/web/hooks/use-file-picker";
 import { ClickToReplaceOverlay } from "./click-to-replace-overlay";
 import { extractUrl } from "./extract-url";
 import type { FieldProps } from "./field-props";
+import { basename, extension } from "./media-filename";
 import { MediaTransformControls } from "./media-transform-controls";
 
 function ExtBadge({ ext }: { ext: string }) {
@@ -24,21 +25,6 @@ function ExtBadge({ ext }: { ext: string }) {
       {ext}
     </span>
   );
-}
-
-function basename(url: string): string {
-  try {
-    const path = new URL(url).pathname;
-    return decodeURIComponent(path.split("/").pop() ?? url);
-  } catch {
-    return url.split("/").pop() ?? url;
-  }
-}
-
-function extension(filename: string): string {
-  const dot = filename.lastIndexOf(".");
-  if (dot < 0 || dot === filename.length - 1) return "";
-  return filename.slice(dot + 1).toLowerCase();
 }
 
 export function FileField({
